The general plan
Oh, a lot has been done here because of the deadline. I was late for the deadline, what can I hide, but the project remained and I had to reconstruct it. And so, in this comment: some temporary prefabs to the procedural generation of the map (a good start, right?), unfinished camera and character control, UI settings, modifiers and modules. And what of this is really ready - nothing. Nothing! But I will improve <3